refactor(test): refactor dataservice unit tests
